Fix on the fix for conversion from Double to Integer.
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@3157 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/be/ibridge/kettle/trans/step/scriptvalues_mod/ScriptValuesMod.java
@@ -332,7 +332,7 @@ public class ScriptValuesMod extends BaseStep implements StepInterface
 								}else if (classType.equalsIgnoreCase("java.lang.Long")){
 									res.setValue( ((Long)result).longValue() );
                                 }else if (classType.equalsIgnoreCase("java.lang.Double")){
-                                    res.setValue( (long)((Double)result).doubleValue() );
+                                    res.setValue( ((Double)result).longValue() );
 								}else if (classType.equalsIgnoreCase("java.lang.String")){
 									res.setValue( (new Long((String)result)).longValue() );									
 								}else if (classType.equalsIgnoreCase("org.mozilla.javascript.Undefined")){
